streamlit-app/requirements.txt
# Core dependencies
streamlit==1.41.1
pillow==11.2.1

# PyTorch (CPU version)
torch==2.7.0
torchvision==0.22.0
